Backend-repo/requirements.txt
# Requirements generated automatically
# Generated on: 2025-09-18T15:54:57.251239

# FastAPI and ASGI server
fastapi==0.104.1
httpx==0.28.1
starlette==0.27.0
uvicorn==0.24.0

# Database
SQLAlchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication and Security
bcrypt==4.1.2
passlib==1.7.4
python-jose==3.3.0
PyJWT==2.10.1

# Utilities and Configuration
aioredis==2.0.1
pydantic==2.11.10
pydantic-settings==2.1.0
pydantic_core==2.33.2
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1

# Other Dependencies
APScheduler==3.11.0
Mako==1.3.10
MarkupSafe==3.0.2
PyYAML==6.0.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
coverage==7.10.6
cryptography==45.0.7
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
httpcore==1.0.9
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
storage3==2.21.1
strenum==0.4.15
supabase-auth==2.21.1
supabase-functions==2.21.1
postgrest==2.21.1
realtime==2.21.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
watchfiles==1.1.0
websockets==15.0.1
supabase==2.21.1
boto3==1.40.74
